LIKA-454: Merge apicatalog_routes to apicatalog
