+++ ansible/deploy-all.yml
@@ -28,7 +28,6 @@
     - postgres
     - postgres-backup
     - { role: ckan-extension, ckanext: ckanext-sentry }
-    - { role: ckan-extension, ckanext: ckanext-apicatalog_routes }
 
     - { role: ckan-extension, ckanext: ckanext-apicatalog }
     - { role: ckan-extension, ckanext: ckanext-scheming }
