+++ ansible/roles/ckan/templates/ckan.ini.j2
@@ -134,9 +134,9 @@ ckanext.forcetranslation.path = i18n
 ckanext.forcetranslation.domain = ckan
 
 
-ckanext.apicatalog_routes.allowed_user_creators = {{ allowed_user_creators | default([]) | join(" ")  }}
-ckanext.apicatalog_routes.allowed_user_editors = {{ allowed_user_editors | default([]) | join(" ")  }}
-ckanext.apicatalog_routes.allowed_member_editors = {{ allowed_member_editors | default([]) | join(" ") }}
+ckanext.apicatalog.allowed_user_creators = {{ allowed_user_creators | default([]) | join(" ")  }}
+ckanext.apicatalog.allowed_user_editors = {{ allowed_user_editors | default([]) | join(" ")  }}
+ckanext.apicatalog.allowed_member_editors = {{ allowed_member_editors | default([]) | join(" ") }}
 
 ckanext.xroad_integration.xroad_gateway_address = http://localhost:9090/rest-adapter-service
 ckanext.xroad_integration.xroad_catalog_address = https://{{ xroad.securityserver.host }}:{{ xroad.securityserver.port }}/r1/{{ xroad.service_id | replace('.', '/') }}{% if xroad.service_name is defined %}/{{ xroad.service_name }}{% endif %}
