+++ ansible/vars/environment-specific/prod.yml
@@ -104,7 +104,6 @@ ckan_plugins:
   - apicatalog_pages
   - pages
   - apicatalog_admin_dashboard
-  - apicatalog_routes
   - contact
   - apicatalog_harvesterstatus
   - harvest
