+++ ansible/vars/environment-specific/test.yml
@@ -99,7 +99,6 @@ ckan_plugins:
   - apicatalog_pages
   - pages
   - apicatalog_admin_dashboard
-  - apicatalog_routes
   - contact
   - apicatalog_harvesterstatus
   - harvest
