+++ ansible/vars/environment-specific/vagrant.yml
@@ -111,7 +111,6 @@ ckan_plugins:
   - apicatalog_pages
   - pages
   - apicatalog_admin_dashboard
-  - apicatalog_routes
   - contact
   - apicatalog_harvesterstatus
   - harvest
