+++ ckanext/ckanext-apicatalog/ckanext/apicatalog/auth.py
@@ -14,7 +14,7 @@ log = logging.getLogger(__name__)
 
 @auth_allow_anonymous_access
 def package_show(context, data_dict):
-    read_only_users = aslist(config.get('ckanext.apicatalog_routes.readonly_users', []))
+    read_only_users = aslist(config.get('ckanext.apicatalog.readonly_users', []))
 
     if context.get('user') and context.get('user') in read_only_users:
         return {'success': True}
@@ -26,7 +26,7 @@ def read_members(context, data_dict):
 
     if 'id' not in data_dict and 'group' not in context:
         data_dict['id'] = c.group_dict['id']
-    read_only_users = aslist(config.get('ckanext.apicatalog_routes.readonly_users', []))
+    read_only_users = aslist(config.get('ckanext.apicatalog.readonly_users', []))
 
     if context.get('user') and context.get('user') in read_only_users:
         return {'success': True}
@@ -38,7 +38,7 @@ def read_members(context, data_dict):
 
 
 def create_user_to_organization(context, data_dict=None):
-    users_allowed_to_create_users = aslist(config.get('ckanext.apicatalog_routes.allowed_user_creators', []))
+    users_allowed_to_create_users = aslist(config.get('ckanext.apicatalog.allowed_user_creators', []))
     if context.get('user') and context.get('user') in users_allowed_to_create_users:
         return {"success": True}
 
@@ -50,7 +50,7 @@ def create_user_to_organization(context, data_dict=None):
 
 @chained_auth_function
 def user_create(next_auth, context, data_dict=None):
-    users_allowed_to_create_users = aslist(config.get('ckanext.apicatalog_routes.allowed_user_editors', []))
+    users_allowed_to_create_users = aslist(config.get('ckanext.apicatalog.allowed_user_editors', []))
     if context.get('user') and context.get('user') in users_allowed_to_create_users:
         return {"success": True}
 
@@ -66,7 +66,7 @@ def user_create(next_auth, context, data_dict=None):
 @chained_auth_function
 @auth_allow_anonymous_access
 def user_update(next_auth, context, data_dict=None):
-    users_allowed_to_create_users = aslist(config.get('ckanext.apicatalog_routes.allowed_user_editors', []))
+    users_allowed_to_create_users = aslist(config.get('ckanext.apicatalog.allowed_user_editors', []))
     if context.get('user_obj'):
         sysadmin_field = context.get('user_obj').sysadmin
     else:
@@ -86,7 +86,7 @@ def user_update(next_auth, context, data_dict=None):
 @chained_auth_function
 @auth_allow_anonymous_access
 def user_show(next_auth, context, data_dict=None):
-    users_allowed_to_create_users = aslist(config.get('ckanext.apicatalog_routes.allowed_user_editors', []))
+    users_allowed_to_create_users = aslist(config.get('ckanext.apicatalog.allowed_user_editors', []))
 
     if context.get('user') and context.get('user') in users_allowed_to_create_users \
             and context['user_obj'].sysadmin is False:
@@ -97,7 +97,7 @@ def user_show(next_auth, context, data_dict=None):
 
 
 def _is_member_editor(context):
-    users_allowed_to_edit_members = aslist(config.get('ckanext.apicatalog_routes.allowed_member_editors', []))
+    users_allowed_to_edit_members = aslist(config.get('ckanext.apicatalog.allowed_member_editors', []))
     return context.get('user') and context.get('user') in users_allowed_to_edit_members
