+++ ckanext/ckanext-apicatalog/ckanext/apicatalog/templates/admin/config.html
@@ -20,5 +20,5 @@
     {{ form.markdown('ckanext.apicatalog.right_column.fi', id='field-ckanext.apicatalog.right_column.fi', label=_('Right column in finnish'), value=data['ckanext.apicatalog.right_column.fi'], error=errors['ckanext.apicatalog.right_column.fi']) }}
     {{ form.markdown('ckanext.apicatalog.right_column.sv', id='field-ckanext.apicatalog.right_column.sv', label=_('Right column in swedish'), value=data['ckanext.apicatalog.right_column.sv'], error=errors['ckanext.apicatalog.right_column.sv']) }}
     {{ form.markdown('ckanext.apicatalog.right_column.en_GB', id='field-ckanext.apicatalog.right_column.en_GB', label=_('Right column in english'), value=data['ckanext.apicatalog.right_column.en_GB'], error=errors['ckanext.apicatalog.right_column.en_GB']) }}
-    {{ form.input('ckanext.apicatalog_routes.readonly_users', id='field-ckanext.apicatalog_routes.readonly_users', label=_('Readonly admins'), value=data['ckanext.apicatalog_routes.readonly_users'], error=errors['ckanext.apicatalog_routes.readonly_users']) }}
+    {{ form.input('ckanext.apicatalog.readonly_users', id='field-ckanext.apicatalog.readonly_users', label=_('Readonly admins'), value=data['ckanext.apicatalog.readonly_users'], error=errors['ckanext.apicatalog.readonly_users']) }}
 {% endblock %}
