+++ ckanext/ckanext-apicatalog/test.ini
@@ -13,7 +13,7 @@ use = config:/usr/lib/ckan/default/src/ckan/test-core.ini
 
 ckan.locale_default = fi
 
-ckan.plugins = apicatalog apicatalog_routes scheming_datasets scheming_organizations fluent
+ckan.plugins = apicatalog scheming_datasets scheming_organizations fluent
 
 scheming.presets = ckanext.scheming:presets.json
                    ckanext.fluent:presets.json
