--- ckanext/ckanext-apicatalog_routes/.coveragerc
@@ -1,5 +1,0 @@
-[report]
-omit =
-    */site-packages/*
-    */python?.?/*
-    ckan/*
