--- ckanext/ckanext-apicatalog_routes/.travis.yml
@@ -1,11 +1,0 @@
-language: python
-python:
-    - "2.6"
-    - "2.7"
-env: PGVERSION=9.1
-install:
-    - bash bin/travis-build.bash
-    - pip install coveralls
-script: sh bin/travis-run.sh
-after_success:
-    - coveralls
