--- ckanext/ckanext-apicatalog_routes/MANIFEST.in
@@ -1,2 +1,0 @@
-include README.rst
-recursive-include ckanext/apicatalog_routes *.html *.json *.js *.less *.css
