--- ckanext/ckanext-apicatalog_routes/dev-requirements.txt
@@ -1,2 +1,0 @@
-pytest-ckan
-pytest-cov
